requirements.txt
bcrypt>=4.0.1
fastapi>=0.96.0
fastapi-another-jwt-auth>=0.1.8
psycopg2-binary>=2.9.6
pydantic>=1.10.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.15
SQLAlchemy-Utils>=0.41.1
uvicorn>=0.22.0
Werkzeug>=2.3.6
